Abstract binary formatter
The current schema validation utilizes a database that is read with BinaryFormatter. This change is the first of a process to remove the need to use BinaryFormatter by separating out the necessary part of the SbdSchemaDatas from what is solely required for the binary database.
